% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{genre_playlist_rec_save}
\alias{genre_playlist_rec_save}
\title{Genre Playlist Recommendation Test Data}
\format{
An object of class dataframe
}
\usage{
data(genre_playlist_rec_save)
}
\description{
Genre Playlist Recommendation Test Data
}
\examples{
genre_playlist_recommendations(test=TRUE)
data(genre_playlist_rec_save)
}
\keyword{dataframe}
