% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genre_playlist_recs.R
\name{genre_playlist_recommendations}
\alias{genre_playlist_recommendations}
\title{This function will allow users to input a playlist to get artist recommendations based on the genres of the artist in their playlist}
\usage{
genre_playlist_recommendations(playlist, access_token, test = FALSE)
}
\arguments{
\item{playlist}{The playlist URI the user wants to input}

\item{access_token}{The API key the user generates from their secret ID and Password so they can access the Spotify API}

\item{test}{If set to TRUE, will return test data of Spotify Top 50 with all default parameters. If FALSE, users can input their own playlist URI}
}
\value{
A dataframe with the columns being the genres of the artists in the playlist you, and each row being a recommended artist from the genre column
}
\description{
This function will allow users to input a playlist to get artist recommendations based on the genres of the artist in their playlist
}
\examples{
genre_playlist_recommendations('37i9dQZEVXbLRQDuF5jeBp', test=TRUE)
# An example using a playlist to get artist recommendations
# based on the genre of the artists in the playlist

# When you want to run your own artists, you will replace 'test = TRUE'
# with your access token.

# If you would like to return test data, please set test = TRUE
}
