% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{genre_rec_saved}
\alias{genre_rec_saved}
\title{Genre Recommendations Test Data}
\format{
An object of class dataframe
}
\usage{
data(genre_rec_saved)
}
\description{
Genre Recommendations Test Data
}
\examples{
genre_recommendations(test=TRUE)
data(genre_rec_saved)
}
\keyword{dataframe}
