% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{max_metrics_data}
\alias{max_metrics_data}
\title{Maximum Metrics Test Data}
\format{
An object of class dataframe
}
\usage{
data(max_metrics_data)
}
\description{
Maximum Metrics Test Data
}
\examples{
max_metrics_summary(test=TRUE)
data(max_metrics_data)
}
\keyword{dataframe}
