% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{min_metrics_data}
\alias{min_metrics_data}
\title{Minimum Metrics Test Data}
\format{
An object of class dataframe
}
\usage{
data(min_metrics_data)
}
\description{
Minimum Metrics Test Data
}
\examples{
min_metrics_summary(test=TRUE)
data(min_metrics_data)
}
\keyword{dataframe}
