% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/min_metric_summary.R
\name{min_metrics_summary}
\alias{min_metrics_summary}
\title{Finding the minimum for a specific music metric}
\usage{
min_metrics_summary(username, URI, access_token, max_or_min, ..., test = FALSE)
}
\arguments{
\item{username}{Character. User input of user's spotify username}

\item{URI}{Playlist identification code. Requires use of function(access_token)}

\item{access_token}{Access token code. Requires use of function(access_token)}

\item{max_or_min}{Identification of desired min or max results (function(max_metric_summary) preferable for max outputs)}

\item{...}{Optional. Additional metrics from spotifyr::package function(get_audio_playlist_features)}

\item{test}{Allows user to use example data in order to see the intended output. Defaulted to test = FALSE description}
}
\value{
A dataframe showing min value of music metric and the associated song from user playlist
}
\description{
Finding the minimum for a specific music metric
}
\examples{
# Here is an example dataframe for when you input the URI of Beyonce.
result <- min_metrics_summary(c('6vWDO969PvNqNYHIOW5v0m'), test = TRUE)
# When you want to run your own playlist, you will remove 'test = TRUE'
# since this argument uses saved data.

# Then replace the URI with the playlist of your choice and use your
# access token as a parameter.
}
