% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{playlist_avg_saved}
\alias{playlist_avg_saved}
\title{Playlist Averages Test Data}
\format{
An object of class dataframe
}
\usage{
data(playlist_avg_saved)
}
\description{
Playlist Averages Test Data
}
\examples{
playlist_averages(test=TRUE)
data(playlist_avg_saved)
}
\keyword{dataframe}
