% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{ranking_save}
\alias{ranking_save}
\title{Ranking Playlist Test Data}
\format{
An object of class dataframe
}
\usage{
data(ranking_save)
}
\description{
Ranking Playlist Test Data
}
\examples{
ranking_function(test=TRUE)
data(ranking_save)
}
\keyword{dataframe}
