% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shuffle_function.R
\name{shuffled_playlist}
\alias{shuffled_playlist}
\title{Shuffling the order of a playlist}
\usage{
shuffled_playlist(username, URI, access_token, ..., test = FALSE)
}
\arguments{
\item{username}{The user's spotify username}

\item{URI}{The playlist URI}

\item{access_token}{The API code the user made using our api_token function}

\item{...}{The type of song information the user wants; the defaults are track.name, track.album.name, track.preview_url}

\item{test}{Allows user to see an example output of saved data}
}
\value{
a dataframe that outputs the shuffled playlist
}
\description{
Shuffling the order of a playlist
}
\examples{
# Here is an example of an output dataframe when you input the Spotify USA top 50 playlist
# When using your own playlist, remember to replace test=TRUE
# This argument is used for a saved data
# Then replace the username, URI and use your access token as a parameter
testing <- shuffled_playlist('spotify', c('37i9dQZEVXbLRQDuF5jeBp'),test= TRUE)

}
