% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{shuffled_saved}
\alias{shuffled_saved}
\title{Shuffle Playlist Test Data}
\format{
An object of class dataframe
}
\usage{
data(shuffled_saved)
}
\description{
Shuffle Playlist Test Data
}
\examples{
shuffled_playlist(test=TRUE)
data(shuffled_saved)
}
\keyword{dataframe}
